refactor(draft-stage): extract metrics derivations to runes module
Moves styleDriftReports / voiceReport / baselineSceneTitle /
sceneTitles plus their freeze-during-streaming caches out of
DraftStage.svelte into draftStageMetrics.svelte.ts, exposed as a
createDraftStageMetrics(store) factory returning $derived-backed
getters. No behavior change.

Part of #34.
